# api/src/main/resources/swagger.yaml
openapi: 3.0.1
info:
    title: Cookbook API
    description: Documentation Cookbook API v1.0
    version: '1.0'
servers:
    - url: http://localhost:8080
      description: Generated server url
paths:
    /ingredient:
        post:
            tags:
                - ingredient
            operationId: createIngredient
            requestBody:
                $ref: '#/components/requestBodies/CreateIngredientCommand'
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/IngredientId'
        get:
            tags:
                - ingredient
            operationId: findAll
            parameters:
                - $ref: '#/components/parameters/IngredientName'
                - $ref: '#/components/parameters/IngredientCategory'
                - $ref: '#/components/parameters/Order'
                - $ref: '#/components/parameters/OrderBy'
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Ingredient'
    /ingredient/{id}:
        get:
            tags:
                - ingredient
            operationId: findById
            parameters:
                - $ref: '#/components/parameters/IngredientId'
            responses:
                "201":
                    description: Created
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Ingredient'
        put:
            tags:
                - ingredient
            operationId: updateIngredient
            parameters:
                - $ref: '#/components/parameters/IngredientId'
            requestBody:
                $ref: '#/components/requestBodies/UpdateIngredientInfoCommand'
            responses:
                '200':
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/IngredientId'
        delete:
            tags:
                - ingredient
            operationId: deleteIngredient
            parameters:
                - $ref: '#/components/parameters/IngredientId'
            responses:
                '204':
                    $ref: '#/components/responses/204'

    /database/refresh:
        get:
            tags:
                - database-controller
            operationId: refreshDatabase
            responses:
                '200':
                    description: OK
components:
    schemas:
        UpdateIngredientInfoCommand:
            type: object
            properties:
                name:
                    $ref: '#/components/schemas/IngredientName'
                description:
                    $ref: '#/components/schemas/IngredientDescription'
                unit:
                    $ref: '#/components/schemas/Unit'
                category:
                    $ref: '#/components/schemas/Category'
                quantityInStock:
                    $ref: '#/components/schemas/QuantityInStock'
        CreateIngredientCommand:
            type: object
            required:
                - name
                - description
                - unit
                - category
            properties:
                name:
                    $ref: '#/components/schemas/IngredientName'
                description:
                    $ref: '#/components/schemas/IngredientDescription'
                unit:
                    $ref: '#/components/schemas/Unit'
                category:
                    $ref: '#/components/schemas/Category'
                quantityInStock:
                    $ref: '#/components/schemas/QuantityInStock'
        Ingredient:
            type: object
            properties:
                id:
                    $ref: '#/components/schemas/IngredientId'
                name:
                    $ref: '#/components/schemas/IngredientName'
                description:
                    $ref: '#/components/schemas/IngredientDescription'
                unit:
                    $ref: '#/components/schemas/Unit'
                category:
                    $ref: '#/components/schemas/Category'
                quantityInStock:
                    $ref: '#/components/schemas/QuantityInStock'
        IngredientId:
            type: integer
            format: int64
            example: 1
        IngredientName:
            type: string
            example: "Tomate"
        IngredientDescription:
            type: string
            example: "Legume rouge"
        QuantityInStock:
            type: integer
            example: 12
        Unit:
            type: string
            example: GRAM
            enum:
                - GRAM
                - LITER
                - PIECE
                - TEASPOON
                - TABLESPOON
                - CUP
                - SLICE
        Category:
            type: string
            example: MEAT
            enum:
                - MEAT
                - VEGETABLE
                - FRUIT
                - DAIRY
                - GRAIN
                - SPICE
                - SWEET
                - BEVERAGE
    parameters:
        IngredientId:
            name: id
            in: path
            required: true
            schema:
                type: integer
                format: int64
        IngredientName:
            name: name
            in: query
            required: false
            schema:
                $ref: '#/components/schemas/IngredientName'
        IngredientCategory:
            name: category
            in: query
            required: false
            schema:
                $ref: '#/components/schemas/Category'
        Order:
            name: order
            in: query
            required: false
            schema:
                type: string
                enum:
                    - asc
                    - desc
        OrderBy:
            name: orderBy
            in: query
            required: false
            schema:
                type: string
                enum:
                    - name
                    - stock
                    - category
                    - unit
    requestBodies:
        UpdateIngredientInfoCommand:
            required: true
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/UpdateIngredientInfoCommand'
        CreateIngredientCommand:
            required: true
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/CreateIngredientCommand'
    responses:
        '200':
            description: Ok
        '201':
            description: Created
        '204':
            description: No content
        '400':
            description: Bad request